% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trading_session.R
\name{add_features.trading_session}
\alias{add_features.trading_session}
\title{Add Features to Trading Session}
\usage{
add_features.trading_session(session, ...)
}
\arguments{
\item{session}{A trading_session object}

\item{...}{Named feature specifications}
}
\value{
Updated trading_session object
}
\description{
Add Features to Trading Session
}
