% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_currency}
\alias{format_currency}
\title{Format Currency}
\usage{
format_currency(x, digits = 2)
}
\arguments{
\item{x}{Numeric value}

\item{digits}{Number of decimal places}
}
\value{
Character string formatted as currency
}
\description{
Format Currency
}
\keyword{internal}
