% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_percentage}
\alias{format_percentage}
\title{Format Percentage}
\usage{
format_percentage(x, digits = 2)
}
\arguments{
\item{x}{Numeric value (e.g., 0.05 for 5\%)}

\item{digits}{Number of decimal places}
}
\value{
Character string formatted as percentage
}
\description{
Format Percentage
}
\keyword{internal}
