% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{freq_to_seconds}
\alias{freq_to_seconds}
\title{Convert Frequency String to Seconds}
\usage{
freq_to_seconds(freq)
}
\arguments{
\item{freq}{Character: frequency string like "1min", "1hour", "daily"}
}
\value{
Numeric: number of seconds
}
\description{
Convert Frequency String to Seconds
}
\keyword{internal}
