% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{has_package}
\alias{has_package}
\title{Check if Package is Available}
\usage{
has_package(pkg)
}
\arguments{
\item{pkg}{Character string of package name}
}
\value{
Logical
}
\description{
Check if Package is Available
}
\keyword{internal}
