% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_tbl.R
\name{is_market_tbl}
\alias{is_market_tbl}
\title{Check if Object is a Market Tibble}
\usage{
is_market_tbl(x)
}
\arguments{
\item{x}{An object to test}
}
\value{
Logical value indicating if object is a \code{market_tbl}
}
\description{
Check if Object is a Market Tibble
}
