% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{plot.backtest_results}
\alias{plot.backtest_results}
\title{Plot Method for Backtest Results}
\usage{
\method{plot}{backtest_results}(x, ...)
}
\arguments{
\item{x}{A \code{backtest_results} object}

\item{...}{Additional arguments}
}
\description{
Plot Method for Backtest Results
}
