% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trading_session.R
\name{plot.trading_session}
\alias{plot.trading_session}
\title{Plot Method for Trading Session}
\usage{
\method{plot}{trading_session}(x, type = "equity", ...)
}
\arguments{
\item{x}{A trading_session object}

\item{type}{Plot type: "equity", "drawdown", "returns", or "all"}

\item{...}{Additional arguments passed to plotting functions}
}
\description{
Plot Method for Trading Session
}
