% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_tbl.R
\name{print.market_tbl}
\alias{print.market_tbl}
\title{Print Method for Market Tibble}
\usage{
\method{print}{market_tbl}(x, ...)
}
\arguments{
\item{x}{A \code{market_tbl} object}

\item{...}{Additional arguments passed to print}
}
\description{
Print Method for Market Tibble
}
