% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trading_session.R
\name{print.trading_session}
\alias{print.trading_session}
\title{Print Method for Trading Session}
\usage{
\method{print}{trading_session}(x, ...)
}
\arguments{
\item{x}{A trading_session object}

\item{...}{Additional arguments}
}
\description{
Print Method for Trading Session
}
