% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trading_session.R
\name{run_backtest.trading_session}
\alias{run_backtest.trading_session}
\title{Run Backtest on Trading Session}
\usage{
run_backtest.trading_session(session, initial_capital = 1e+05, ...)
}
\arguments{
\item{session}{A trading_session object}

\item{initial_capital}{Initial capital (default: 100000)}

\item{...}{Additional backtest parameters}
}
\value{
Updated trading_session object with backtest results
}
\description{
Run Backtest on Trading Session
}
