% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sma}
\alias{sma}
\title{Calculate Simple Moving Average}
\usage{
sma(x, n = 20)
}
\arguments{
\item{x}{Numeric vector}

\item{n}{Window size}
}
\value{
Numeric vector of same length with SMA values
}
\description{
Helper function to calculate SMA for examples and testing
}
\examples{
prices <- c(100, 102, 101, 103, 105, 104, 106)
sma(prices, n = 3)
}
