% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtest.R
\name{summary.backtest_results}
\alias{summary.backtest_results}
\title{Summary Method for Backtest Results}
\usage{
\method{summary}{backtest_results}(object, ...)
}
\arguments{
\item{object}{A \code{backtest_results} object}

\item{...}{Additional arguments}
}
\description{
Summary Method for Backtest Results
}
