% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trading_session.R
\name{summary.trading_session}
\alias{summary.trading_session}
\title{Summary Method for Trading Session}
\usage{
\method{summary}{trading_session}(object, ...)
}
\arguments{
\item{object}{A trading_session object}

\item{...}{Additional arguments}
}
\description{
Summary Method for Trading Session
}
